pivot_based_eccv2018/misc/expander/requirements.txt
nltk==3.2.5
PyYAML==3.12
six==1.11.0
